Reoragnise the `usart_spi` into its own module. + Docs
